notebooks/pyproject.toml
[project]
name = "ssondo-notebooks"
version = "0.1.0"
description = "Notebooks for S-SONDO evaluation"
requires-python = ">=3.10"
dependencies = [
    "ssondo>=0.1.0",
    "soundfile>=0.12.1",
    "librosa>=0.11.0",
    "datasets>=3.0,<4.0",
    "scikit-learn>=1.4",
    "matplotlib>=3.8",
    "tqdm>=4.66",
    "umap-learn>=0.5",
    "ipykernel>=6.29",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
